Moved Store Into Redux Folder
