+++ src/components/App/Providers.tsx
@@ -4,7 +4,7 @@ import { ApolloProvider } from 'react-apollo';
 import { Provider } from 'react-redux';
 import { HashRouter } from 'react-router-dom';
 
-import store from '../../store';
+import store from '../../redux/store';
 import client from '../../graphql/client';
 
 export const Providers = ({ children }: { children: React.ReactNode }) => (
