+++ src/stories/Queue.tsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 
 import Queue from '../containers/Queue';
 import { Providers } from '../components/App/Providers';
-import store from '../store';
+import store from '../redux/store';
 import { populateQueue } from './populateQueue';
 
 store.dispatch(populateQueue());
